create a settings page and edits routes
