+++ front/src/App.js
@@ -4,7 +4,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 //Components
 import { HomePage } from './pages/HomePage';
 import { AuthPage } from './pages/AuthPage';
-import { AmenitiesPage } from './pages/AmenitiesPage';
+import { SettingsPage } from './pages/SettingsPage';
 import { Google } from "./pages/GoolePage";
 
 // HOCs
@@ -17,7 +17,7 @@ export const App = withAuth(() => {
                 <Switch>
                     <Route path="/" exact component={HomePage} />
                     <Route path="/auth" component={AuthPage} />
-                    <Route path="/amenities" component={AmenitiesPage} />
+                    <Route path="/settings" component={SettingsPage} />
                     <Route path="/google/login" exact component={Google} />
                 </Switch>
             </Router>
